# docs/swagger.yaml
basePath: /api/telemetry
definitions:
  github_com_seongpil0948_otel-kafka-pg_modules_api_dto.ErrorInfo:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_api_dto.LogsResponse:
    properties:
      logs:
        items:
          $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_log_domain.LogItem'
        type: array
      pagination:
        $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Pagination'
      services:
        items:
          type: string
        type: array
      severities:
        items:
          type: string
        type: array
      timeRange:
        $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.TimeRange'
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Pagination:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.ErrorInfo'
      success:
        type: boolean
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_api_dto.ServiceMetric:
    properties:
      avgLatency:
        type: number
      errorCount:
        type: integer
      errorRate:
        type: number
      name:
        type: string
      p95Latency:
        type: number
      p99Latency:
        type: number
      requestCount:
        type: integer
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_api_dto.ServiceMetricsResponse:
    properties:
      avgLatency:
        type: number
      errorPercentage:
        type: number
      services:
        items:
          $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.ServiceMetric'
        type: array
      timeRange:
        $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.TimeRange'
      totalErrors:
        type: integer
      totalRequests:
        type: integer
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_api_dto.TimeRange:
    properties:
      endTime:
        type: integer
      startTime:
        type: integer
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_api_dto.TraceDetailResponse:
    properties:
      relatedLogs:
        items:
          $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_log_domain.LogItem'
        type: array
      trace:
        $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_trace_domain.Trace'
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_api_dto.TracesResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Pagination'
      services:
        items:
          type: string
        type: array
      timeRange:
        $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.TimeRange'
      totalDuration:
        type: integer
      traces:
        items:
          $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_trace_domain.TraceItem'
        type: array
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_log_domain.LogItem:
    properties:
      attributes:
        additionalProperties: true
        type: object
      id:
        type: string
      message:
        type: string
      serviceName:
        type: string
      severity:
        type: string
      spanId:
        type: string
      timestamp:
        type: integer
      traceId:
        type: string
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_trace_domain.ServiceInfo:
    properties:
      avgLatency:
        type: number
      count:
        type: integer
      errorCount:
        type: integer
      errorRate:
        type: number
      name:
        type: string
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_trace_domain.ServiceListResult:
    properties:
      services:
        items:
          $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_trace_domain.ServiceInfo'
        type: array
      took:
        type: integer
      total:
        type: integer
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_trace_domain.Span:
    properties:
      attributes:
        additionalProperties: true
        type: object
      duration:
        type: number
      endTime:
        type: integer
      id:
        type: string
      name:
        type: string
      parentSpanId:
        type: string
      serviceName:
        type: string
      spanId:
        type: string
      startTime:
        type: integer
      status:
        type: string
      traceId:
        type: string
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_trace_domain.Trace:
    properties:
      endTime:
        type: integer
      services:
        items:
          type: string
        type: array
      spans:
        items:
          $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_trace_domain.Span'
        type: array
      startTime:
        type: integer
      total:
        type: integer
      traceId:
        type: string
    type: object
  github_com_seongpil0948_otel-kafka-pg_modules_trace_domain.TraceItem:
    properties:
      attributes:
        additionalProperties: true
        type: object
      duration:
        type: number
      endTime:
        type: integer
      id:
        type: string
      name:
        type: string
      parentSpanId:
        type: string
      serviceName:
        type: string
      spanId:
        type: string
      startTime:
        type: integer
      status:
        type: string
      traceId:
        type: string
    type: object
externalDocs:
  description: OpenTelemetry 설명 문서
  url: https://opentelemetry.io/docs/
host: localhost:8080
info:
  contact:
    email: support@example.org
    name: API Support Team
    url: http://example.org/support
  description: OpenTelemetry 텔레메트리 데이터를 위한 RESTful API 서비스
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OpenTelemetry API
  version: 1.0.0
paths:
  /logs:
    get:
      consumes:
      - application/json
      description: 필터 조건에 맞는 로그 목록을 조회합니다
      parameters:
      - description: 시작 시간 (밀리초 타임스탬프)
        in: query
        name: startTime
        type: integer
      - description: 종료 시간 (밀리초 타임스탬프)
        in: query
        name: endTime
        type: integer
      - collectionFormat: csv
        description: 서비스 이름 목록
        in: query
        items:
          type: string
        name: serviceNames
        type: array
      - description: 심각도 (INFO, WARN, ERROR, FATAL 등)
        in: query
        name: severity
        type: string
      - description: 트레이스 ID가 있는 로그만 필터링
        in: query
        name: hasTrace
        type: boolean
      - description: 검색어
        in: query
        name: query
        type: string
      - default: 20
        description: 한 페이지당 항목 수
        in: query
        name: limit
        type: integer
      - default: 0
        description: 오프셋
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.LogsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
      summary: 로그 목록 조회
      tags:
      - logs
  /logs/summary:
    get:
      consumes:
      - application/json
      description: 특정 기간의 로그에 대한 요약 정보를 조회합니다
      parameters:
      - description: 시작 시간 (밀리초 타임스탬프)
        in: query
        name: startTime
        type: integer
      - description: 종료 시간 (밀리초 타임스탬프)
        in: query
        name: endTime
        type: integer
      - description: 서비스 이름(선택적)
        in: query
        name: serviceName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
      summary: 로그 요약 정보 조회
      tags:
      - logs
  /logs/trace/{traceId}:
    get:
      consumes:
      - application/json
      description: 특정 트레이스 ID와 관련된 로그를 조회합니다
      parameters:
      - description: Trace ID
        in: path
        name: traceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.LogsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
      summary: 트레이스 ID로 관련 로그 조회
      tags:
      - logs
  /metrics/services:
    get:
      consumes:
      - application/json
      description: 서비스별 성능 지표를 조회합니다
      parameters:
      - description: 시작 시간 (밀리초 타임스탬프)
        in: query
        name: startTime
        type: integer
      - description: 종료 시간 (밀리초 타임스탬프)
        in: query
        name: endTime
        type: integer
      - description: 서비스 이름(선택적)
        in: query
        name: serviceName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.ServiceMetricsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
      summary: 서비스 지표 조회
      tags:
      - metrics
  /traces:
    get:
      consumes:
      - application/json
      description: 필터 조건에 맞는 트레이스 목록을 조회합니다
      parameters:
      - description: 시작 시간 (밀리초 타임스탬프)
        in: query
        name: startTime
        type: integer
      - description: 종료 시간 (밀리초 타임스탬프)
        in: query
        name: endTime
        type: integer
      - collectionFormat: csv
        description: 서비스 이름 목록
        in: query
        items:
          type: string
        name: serviceNames
        type: array
      - description: 상태 (OK, ERROR 등)
        in: query
        name: status
        type: string
      - description: 최소 지속 시간 (밀리초)
        in: query
        name: minDuration
        type: integer
      - description: 최대 지속 시간 (밀리초)
        in: query
        name: maxDuration
        type: integer
      - description: 검색어
        in: query
        name: query
        type: string
      - default: 20
        description: 한 페이지당 항목 수
        in: query
        name: limit
        type: integer
      - default: 0
        description: 오프셋
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.TracesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
      summary: 트레이스 목록 조회
      tags:
      - traces
  /traces/{traceId}:
    get:
      consumes:
      - application/json
      description: 특정 트레이스 ID에 대한 상세 정보를 조회합니다
      parameters:
      - description: Trace ID
        in: path
        name: traceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.TraceDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
      summary: 트레이스 ID로 상세 정보 조회
      tags:
      - traces
  /traces/services:
    get:
      consumes:
      - application/json
      description: 모든 서비스 목록과 기본 통계 정보를 조회합니다
      parameters:
      - description: 시작 시간 (밀리초 타임스탬프)
        in: query
        name: startTime
        type: integer
      - description: 종료 시간 (밀리초 타임스탬프)
        in: query
        name: endTime
        type: integer
      - description: 서비스명 필터링 검색어
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_trace_domain.ServiceListResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_seongpil0948_otel-kafka-pg_modules_api_dto.Response'
      summary: 서비스 목록 조회
      tags:
      - traces
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"